% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateScores.R
\name{calculateScores}
\alias{calculateScores}
\title{Calculate scores}
\usage{
calculateScores()
}
\value{
Table of z-scores.
}
\description{
This function calculates z-scores from normalized coverage matrix.
}
\examples{
calculateScores()
}
