% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineEnvironment.R
\name{defineFiles}
\alias{defineFiles}
\title{Define files}
\usage{
defineFiles(FILE = FILE)
}
\arguments{
\item{FILE}{name of *.bcmatrix.xlsx in current working directory}
}
\description{
This function defines environmental files.
}
\examples{
defineFiles("/path/to/bcmatric.xlsx")
}
