% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeData.R
\name{normalizeGenes}
\alias{normalizeGenes}
\title{Normalize genes}
\usage{
normalizeGenes()
}
\description{
This function extracts the normalized gene coverage from the dataframe.
}
\examples{
normalizeGenes()
}
