% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refValues.R
\docType{data}
\name{REFVALUES}
\alias{REFVALUES}
\title{Reference values}
\format{An object of class \code{"datamatrix"}.}
\usage{
data("refValues", package="CoveRageAnalysis")
}
\description{
Reference values
}
\keyword{reference}
