% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportResults.R
\name{reportAmplifications}
\alias{reportAmplifications}
\title{Report focal amplification}
\usage{
reportAmplifications()
}
\value{
List of samples with focal gene amplification.
}
\description{
This function prints summary of focal amplification z-scores.
}
\examples{
reportAmplifications()
}
