% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportResults.R
\name{reportV3}
\alias{reportV3}
\title{Report EGFRvIII}
\usage{
reportV3()
}
\value{
List of samples with EGFRvIII
}
\description{
This function prints summary of EGFRvIII intragenic deletion z-scores.
}
\examples{
reportV3()
}
