% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFiles.R
\name{writeFiles}
\alias{writeFiles}
\title{Write results files}
\usage{
writeFiles()
}
\value{
The directory containing exported files.
}
\description{
This function writes comma-separated files to the project directory.
}
\examples{
writeFiles()
}
